# api/openapi.yaml
openapi: 3.1.0
info:
  title: Domestique API
  version: 2.0.0
  description: |
    The API contract served by Domestique. JSON property names are camelCase.
    All routes except `GET /healthz` require the browser session cookie;
    state-changing routes also require the same-origin `Origin` header. The
    sign-in routes under `/auth/` are browser navigation rather than API
    operations, and are not described here.
servers:
  - url: /
security:
  - session: []
...
paths:
...
  /v1/routes:
    get:
      tags:
        - browser-api
      operationId: getRoutes
      responses:
        "200":
          description: Stored route inventory without geometry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "503":
          $ref: '#/components/responses/Unavailable'
  /v1/providers/{provider}/sourceRoutes/{sourceRouteId}/routes/{stageOrder}:
    get:
      tags:
        - browser-api
      operationId: getRoute
      parameters:
        - $ref: '#/components/parameters/Provider'
        - $ref: '#/components/parameters/SourceRouteId'
        - $ref: '#/components/parameters/StageOrder'
      responses:
        "200":
          description: Stored route metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "503":
          $ref: '#/components/responses/Unavailable'
...
  /v1/webui/config:
    get:
      tags:
        - browser-api
      operationId: getWebUIConfig
      responses:
        "200":
          description: Browser-only runtime configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebUIConfig'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-domestique_session
      description: Browser session cookie issued by the sign-in flow.
    browserOrigin:
      type: apiKey
      in: header
      name: Origin
      description: Exact browser origin configured for this service. Required by every operation that starts a run or writes state, and modelled as a security scheme rather than a parameter because a request without it is refused rather than misread.
  parameters:
...
    Provider:
      name: provider
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    SourceRouteId:
      name: sourceRouteId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    StageOrder:
      name: stageOrder
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
...
    InvalidRequest:
      description: Request is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid browser session.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Session identity or origin is not permitted.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unavailable:
      description: Local state is unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
...
    RouteValidation:
      type: object
      additionalProperties: false
      required:
        - biasPercent
        - maePercent
        - p90Percent
        - evaluatedRides
      properties:
        biasPercent:
          type: number
          format: double
        maePercent:
          type: number
          format: double
        p90Percent:
          type: number
          format: double
        evaluatedRides:
          type: integer
          minimum: 0
    Route:
      type: object
      additionalProperties: false
      required:
        - provider
        - sourceRouteId
        - stageOrder
        - title
        - sourceRouteName
        - routeName
        - sourceRevision
        - contentHash
        - distanceMetres
        - ascentMetres
        - descentMetres
        - maxGradientPercent
        - pointCount
      properties:
        provider:
          type: string
        sourceRouteId:
          type: integer
          format: int64
        stageOrder:
          type: integer
        title:
          type: string
        sourceRouteName:
          type: string
        routeName:
          type: string
        sourceRevision:
          type: string
        contentHash:
          type: string
        distanceMetres:
          type: number
          format: double
        ascentMetres:
          type: number
          format: double
        descentMetres:
          type: number
          format: double
        maxGradientPercent:
          type: number
          format: double
        pointCount:
          type: integer
        movingSeconds:
          type: number
          format: double
        validation:
          $ref: '#/components/schemas/RouteValidation'
    RouteList:
      type: object
      additionalProperties: false
      required:
        - routes
      properties:
        routes:
          type: array
          items:
            $ref: '#/components/schemas/Route'
...
    BrowserBasemap:
      type: object
      additionalProperties: false
      required:
        - name
        - styleUrl
      properties:
        name:
          type: string
          minLength: 1
        styleUrl:
          type: string
          format: uri
        styleUrlDark:
          type: string
          format: uri
        darkCartography:
          type: boolean
    WebUIConfig:
      type: object
      additionalProperties: false
      required:
        - basemaps
        - identity
      properties:
        basemaps:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/BrowserBasemap'
        sourceBaseUrls:
          $ref: '#/components/schemas/SourceBaseUrls'
        identity:
          $ref: '#/components/schemas/BrowserIdentity'
    BrowserIdentity:
      type: object
      additionalProperties: false
      description: >-
        Who the gate let through. It names the reader rather than identifying
        them: it is how a session can be seen to be the one intended.
      required:
        - display
        - admin
      properties:
        display:
          type: string
        admin:
          type: boolean
          description: >-
            Whether this subject holds cross-subject rights, asserted by the
            Auth0 Action that decided it as well as whether this subject may
            sign in at all.
    SourceBaseUrls:
      type: object
      additionalProperties: false
      properties:
        veloplanner:
          type: string
          format: uri
        komoot:
          type: string
          format: uri
...